fix memory safety violation / use after free
before DLexer would take slices from the mutable/const input and just
cast(string) them, which violates expectation from the caller that they
can modify the memory after the memory finished. This duplicates it on
creation of the DLexer struct, fixing corruptions on freed memory or
other modifications.

More concretely this fixes a segfault in DCD trying to parse a large
module with comments, causing libdparse to call the .comment method in
token trivia, which attempts to read the corrupted text content of the
comment tokens to check for ddoc. (test file was
Pure-D/workspace-d/source/workspaced/com/dcd.d)